requirements.txt
mamba
ffmpeg-python
Pillow
numpy
pandas
matplotlib
ipython
ipykernel
ipywidgets
nbformat
tqdm
pytube
pydantic
minio
requests
python-dotenv
opencv-python
yt-dlp
pytorch_lightning
torchvision==0.15.2
torch==2.0.1
torchinfo==1.8.0
torchaudio